NB1: Do not extract/flash hidden partition
/hidden stores Nokia USB drivers, used in stock and it will mount hidden
as a CD-ROM drive for us to install the drivers on Windows.

Retrofit A/B needs more than one partition unlike A-only and trying to
retrofit only system partition will cause the build to fail. Group hidden
with system allows us to pass lpmake checks when creating super image.

That will do as a workaround for now; until we found a feasible solution for this.

Signed-off-by: Tuan Anh <tuan73176@gmail.com>
Change-Id: I7d46b65478ab726179f899bb64649d4e7247d1b2